--- src/components/logo/classes.ts
@@ -1,7 +1,0 @@
-import { createClasses } from 'src/theme/create-classes';
-
-// ----------------------------------------------------------------------
-
-export const logoClasses = {
-  root: createClasses('logo__root'),
-};
