--- src/components/logo/index.ts
@@ -1,3 +1,0 @@
-export * from './logo';
-
-export * from './classes';
