+++ src/layouts/auth/layout.tsx
@@ -8,7 +8,8 @@ import Alert from '@mui/material/Alert';
 
 import { RouterLink } from 'src/routes/components';
 
-//import { Logo } from 'src/components/logo';
+import { LogoConagua } from 'src/components/logo/logo-conagua';
+
 import { AuthContent } from './content';
 import { MainSection } from '../core/main-section';
 import { LayoutSection } from '../core/layout-section';
@@ -48,12 +49,12 @@ export function AuthLayout({
           This is an info Alert.
         </Alert>
       ),
-     // leftArea: (
-       // <>
-         // {/** @slot Logo */}
-         // <Logo />
-       // </>
-      //),
+        leftArea: (
+        <>
+          {/** @slot Logo */}
+          <LogoConagua />
+        </>
+      ),
       rightArea: (
         <Box sx={{ display: 'flex', alignItems: 'center', gap: { xs: 1, sm: 1.5 } }}>
           {/** @slot Help link */}
