+++ src/layouts/components/notifications-popover.tsx
@@ -1,3 +1,4 @@
+//MENU DE NOTIFICACIONES
 import type { IconButtonProps } from '@mui/material/IconButton';
 
 import { useState, useCallback } from 'react';
