+++ src/layouts/dashboard/layout.tsx
@@ -7,22 +7,21 @@ import Box from '@mui/material/Box';
 import Alert from '@mui/material/Alert';
 import { useTheme } from '@mui/material/styles';
 
-import { _langs, _notifications } from 'src/_mock';
-
+//import { _langs, _notifications } from 'src/_mock';
 import { NavMobile, NavDesktop } from './nav';
 import { layoutClasses } from '../core/classes';
 import { _account } from '../nav-config-account';
 import { dashboardLayoutVars } from './css-vars';
 import { navData } from '../nav-config-dashboard';
 import { MainSection } from '../core/main-section';
-import { Searchbar } from '../components/searchbar';
+//import { Searchbar } from '../components/searchbar';
 import { _workspaces } from '../nav-config-workspace';
 import { MenuButton } from '../components/menu-button';
 import { HeaderSection } from '../core/header-section';
 import { LayoutSection } from '../core/layout-section';
 import { AccountPopover } from '../components/account-popover';
-import { LanguagePopover } from '../components/language-popover';
-import { NotificationsPopover } from '../components/notifications-popover';
+//import { LanguagePopover } from '../components/language-popover';
+//import { NotificationsPopover } from '../components/notifications-popover';
 
 import type { MainSectionProps } from '../core/main-section';
 import type { HeaderSectionProps } from '../core/header-section';
@@ -78,13 +77,13 @@ export function DashboardLayout({
       rightArea: (
         <Box sx={{ display: 'flex', alignItems: 'center', gap: { xs: 0, sm: 0.75 } }}>
           {/** @slot Searchbar */}
-          <Searchbar />
+          {/*<Searchbar />*/}
 
           {/** @slot Language popover */}
-          <LanguagePopover data={_langs} />
+          {/*<LanguagePopover data={_langs} />*/}
 
           {/** @slot Notifications popover */}
-          <NotificationsPopover data={_notifications} />
+          {/*<NotificationsPopover data={_notifications} />*/}
 
           {/** @slot Account drawer */}
           <AccountPopover data={_account} />
