+++ src/sections/overview/analytics-website-visits.tsx
@@ -36,7 +36,7 @@ export function AnalyticsWebsiteVisits({ title, subheader, chart, sx, ...other }
     stroke: { width: 2, colors: ['transparent'] },
     xaxis: { categories: chart.categories },
     legend: { show: true },
-    tooltip: { y: { formatter: (value: number) => `${value} visits` } },
+    tooltip: { y: { formatter: (value: number) => `${value} formatos` } },
     ...chart.options,
   });
